Fix critical bug in zap.
As it turns out, there were some changes in the julienschmidt/httprouter
project that break certain assumptions made about wildcard handling.
Most likely, this is related to commit 6f3f391, and manifested in issues
183 and 172.

Zap unit tests did not catch this since we use a HTTP recorder rather
than the actual framework.

For now, adding a small hack that registers the index handler as the
notfound handler.